chore: move validation code into block header
